+++ client/projects/client/src/app/modules/main/components/layout/sidebar/sidebar.component.ts
@@ -1,9 +1,14 @@
-import { Component } from "@angular/core";
+import { Component, OnInit } from "@angular/core";
+import { version } from '../../../../../../../../../package.json';
 
 @Component({
     selector: 'app-sidebar',
     templateUrl: './sidebar.component.html',
 })
-export class SidebarComponent {
-    
+export class SidebarComponent implements OnInit {
+    version: string
+
+    ngOnInit(): void {
+        this.version = version
+    }
 }