+++ client/projects/lib/src/lib/types/xmpp.ts
@@ -1,3 +1,6 @@
+export interface IXmppService {
+    connect(id: string, token: string, host: string): void
+}
 
 export interface XmppUserToken {
     token: string